Add a Clock option to our LRU cache (#4593)
<!-- Describe what has changed in this PR -->
**What changed?**
I added a Clock to our LRU cache, and I modified our tests to use that
instead of time.Sleep.

<!-- Tell your future self why have you made these changes -->
**Why?**
To make our tests faster, and to support time-skipping in the future

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
A flaky test here would probably make it hang forever. If someone uses
clock.Sleep instead of clock.Advance in the future, that can happen.

<!-- Is this PR a hotfix candidate or require that a notification be
sent to the broader community? (Yes/No) -->
**Is hotfix candidate?**